refactor: refactored MovieScreen split it to multiple sections.
